ui/backend/pyproject.toml
[project]
name = "umbrella-ui-backend"
version = "0.1.0"
description = "Umbrella UI backend — FastAPI server for compliance review"
requires-python = ">=3.11"
dependencies = [
    "umbrella-connector-framework",
    "fastapi>=0.115",
    "uvicorn>=0.32",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "bcrypt<4.0.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "structlog>=24.0",
    "httpx>=0.27",
    "elasticsearch[async]>=8.0,<9.0",
    "boto3>=1.35",
]

[project.scripts]
umbrella-ui = "umbrella_ui.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["umbrella_ui"]

[tool.uv.sources]
umbrella-connector-framework = { path = "../../connectors/connector-framework", editable = true }

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
